[Rename] rename notification and report
