+++ src/main/java/com/example/breadbook/domain/member/model/MemberDto.java
@@ -1,6 +1,5 @@
 package com.example.breadbook.domain.member.model;
 
-import com.example.breadbook.Notification.model.Notification;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,8 +8,6 @@ import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 public class MemberDto {
     @Getter
