+++ src/main/java/com/example/breadbook/domain/notification/NotificationRepository.java
@@ -1,6 +1,6 @@
-package com.example.breadbook.Notification;
+package com.example.breadbook.domain.notification;
 
-import com.example.breadbook.Notification.model.Notification;
+import com.example.breadbook.domain.notification.model.Notification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
