+++ src/main/java/com/example/breadbook/domain/notification/NotificationService.java
@@ -1,8 +1,8 @@
-package com.example.breadbook.Notification;
+package com.example.breadbook.domain.notification;
 
 
-import com.example.breadbook.Notification.model.Notification;
-import com.example.breadbook.Notification.model.NotificationDto;
+import com.example.breadbook.domain.notification.model.Notification;
+import com.example.breadbook.domain.notification.model.NotificationDto;
 import com.example.breadbook.domain.member.model.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
