+++ src/main/java/com/example/breadbook/domain/report/ReportController.java
@@ -1,6 +1,6 @@
-package com.example.breadbook.Report;
+package com.example.breadbook.domain.report;
 
-import com.example.breadbook.Report.model.ReportDto;
+import com.example.breadbook.domain.report.model.ReportDto;
 import com.example.breadbook.domain.member.model.Member;
 import com.example.breadbook.domain.product.model.Product;
 import lombok.RequiredArgsConstructor;
