+++ src/main/java/com/example/breadbook/domain/report/ReportRepository.java
@@ -1,6 +1,6 @@
-package com.example.breadbook.Report;
+package com.example.breadbook.domain.report;
 
-import com.example.breadbook.Report.model.Report;
+import com.example.breadbook.domain.report.model.Report;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReportRepository extends JpaRepository<Report, Long> {
